ImpactWidget refactor
