+++ components/main/blocks/ImpactWidget/ImpactWidget.module.scss
@@ -37,7 +37,7 @@
   border-bottom: 1px solid var(--primary);
   background: var(--secondary);
   position: relative;
-  width: 140px;
+  width: 170px;
   text-align: right;
   padding-right: 34px;
   height: 33px;
