+++ components/main/blocks/ImpactWidget/ImpactWidget.tsx
@@ -1,8 +1,9 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import AnimateHeight from "react-animate-height";
 import styles from "./ImpactWidget.module.scss";
 import { Links } from "../Links/Links";
 import { Spinner } from "../../../shared/components/Spinner/Spinner";
+import { thousandize } from "../../../../util/formatting";
 
 export type Intervention = {
   title: string;
@@ -11,15 +12,51 @@ export type Intervention = {
   outputStringTemplate: string;
 };
 
-export const ImpactWidget: React.FC<{
-  title: string;
-  defaultSum: number;
-  interventions: Intervention[];
-  buttonText: string;
-}> = ({ title, defaultSum, interventions, buttonText }) => {
-  const [sum, setSum] = useState(defaultSum);
-  const [selectedIntervention, setSelectedIntervention] = useState<string>(interventions[0].title);
+export interface ImpactWidgetProps {
+  data: any;
+}
+
+export const ImpactWidget: React.FC<ImpactWidgetProps> = ({ data }) => {
+  const [sum, setSum] = useState(400);
   const [contextExpanded, setContextExpanded] = useState(false);
+  const [interventionCosts, setInterventionCosts] = useState<Map<string, number>>(new Map());
+
+  const interventionWidget = data.result.frontpage[0].intervention_widget;
+  const interventions= interventionWidget.interventions.map((i: any) => ({
+    title: i.title,
+    pricePerOutput: interventionCosts.get(i.abbreviation),
+    outputStringTemplate: i.template_string,
+    organizationName: i.organization_name,
+  }))
+  
+  const [selectedIntervention, setSelectedIntervention] = useState<string>(interventions[0].title);
+
+  useEffect(() => {
+    const interventions = interventionWidget.interventions;
+    const url = `https://impact.gieffektivt.no/api/evaluations?${interventions
+      .map((i: any) => `charity_abbreviation=${i.abbreviation}&`)
+      .join("")}currency=NOK`;
+    console.log(url);
+    fetch(url).then((res) => {
+      res.json().then((data) => {
+        const costs = new Map();
+        const evaluations = data.evaluations;
+        interventions.forEach((i: any) => {
+          // For each intervention, filter the evaluations for a given charity
+          const filtered = evaluations.filter(
+            (e: any) => e.charity.abbreviation === i.abbreviation,
+          );
+          // Then order the list to get the most recent
+          const ordered = filtered.sort((a: any, b: any) => a.start_year - b.start_year);
+          // Get the most recent evaluation
+          const evaluation = ordered[0];
+          // Set the cost to the most recent evaluation converted cost (cost in NOK per output)
+          costs.set(i.abbreviation, evaluation.converted_cost_per_output);
+        });
+        setInterventionCosts(costs);
+      });
+    });
+  }, [interventionWidget.interventions]);
 
   const currentIntervention = interventions.find(
     (i) => i.title === selectedIntervention,
@@ -38,7 +75,7 @@ export const ImpactWidget: React.FC<{
   return (
     <div className={styles.wrapper}>
       <div className={styles.header}>
-        <h3>{title}</h3>
+        <h3>{interventionWidget.title}</h3>
       </div>
       <div className={styles.grid}>
         <div className={styles.input}>
@@ -48,7 +85,9 @@ export const ImpactWidget: React.FC<{
               <input
                 type="tel"
                 value={sum}
-                onChange={(e) => setSum(parseInt(e.target.value) || 0)}
+                onChange={(e) => {
+                  if(e.target.value.length <= 9) setSum(parseInt(e.target.value) || 0)
+                }}
                 name="sum"
               />
             </div>
@@ -78,10 +117,10 @@ export const ImpactWidget: React.FC<{
             <>
               <span className="detailheader">{currentIntervention.organizationName}</span>
               <div className={styles.paragraphWrapper}>
-                <span className={styles.paragraphNumber}>{outputString}</span>
+                <span className={styles.paragraphNumber}>{thousandize(parseInt(outputString))}</span>
                 <div className={styles.explanatory}>
                   <p>
-                    <span className={styles.innerParagraphNumber}>{outputString}&nbsp;</span>
+                    <span className={styles.innerParagraphNumber}>{outputString}47&nbsp;</span>
                     {currentIntervention.outputStringTemplate}
                   </p>
                   <span
